modif strlen - type size_t pour index

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: doberes <doberes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 14:56:30 by doberes           #+#    #+#             */
-/*   Updated: 2024/11/09 13:25:48 by doberes          ###   ########.fr       */
+/*   Updated: 2024/12/02 10:00:10 by doberes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ gcc -Wall -Wextra -Werror ft_strlen.c
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
